refactor(table): group shared support into src/shared for composite archetype
