+++ components/table/src/components/TableBody/render-helper.ts
@@ -4,8 +4,8 @@ import { computed, h, inject } from 'vue'
 import { merge } from 'lodash-unified'
 import { useNamespace } from '@flash-global66/g-utils'
 import { isBoolean, isPropAbsent } from '@flash-global66/g-utils'
-import { getRowIdentity } from '../../utils/table.util'
-import { TABLE_INJECTION_KEY } from '../../constants/token.constant'
+import { getRowIdentity } from '../../shared/utils/table.util'
+import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
 import useEvents from './events-helper'
 import useStyles from './styles-helper'
 import TdWrapper from './td-wrapper.vue'
