+++ components/table/src/components/TableColumn/index.ts
@@ -13,8 +13,8 @@ import {
 } from 'vue'
 import GCheckbox from '@flash-global66/g-checkbox'
 import { isArray, isString, isUndefined } from '@flash-global66/g-utils'
-import { cellStarts } from '../../constants/renderer.constant'
-import { compose, mergeOptions } from '../../utils/table.util'
+import { cellStarts } from '../../shared/constants/renderer.constant'
+import { compose, mergeOptions } from '../../shared/utils/table.util'
 import useWatcher from './watcher-helper'
 import useRender from './render-helper'
 import defaultProps from './defaults'
