+++ components/table/src/components/TableFooter/index.ts
@@ -2,10 +2,10 @@
 // @ts-nocheck
 import { defineComponent, h, inject } from 'vue'
 import { useNamespace } from '@flash-global66/g-utils'
-import useLayoutObserver from '../../composables/useLayoutObserver'
-import { TABLE_INJECTION_KEY } from '../../constants/token.constant'
+import useLayoutObserver from '../../shared/composables/useLayoutObserver'
+import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
 import useStyle from './style-helper'
-import type { Store } from '../../store'
+import type { Store } from '../../shared/store'
 
 import type { PropType } from 'vue'
 import type { DefaultRow, Sort, SummaryMethod } from '../../Table/defaults'
