+++ components/table/src/components/TableHeader/style.helper.ts
@@ -6,8 +6,8 @@ import {
   ensurePosition,
   getFixedColumnOffset,
   getFixedColumnsClass,
-} from '../../utils/table.util'
-import { TABLE_INJECTION_KEY } from '../../constants/token.constant'
+} from '../../shared/utils/table.util'
+import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
 import type { TableColumnCtx } from '../TableColumn/defaults'
 import type { TableHeaderProps } from '.'
