+++ components/table/src/components/TableHeader/utils-helper.ts
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { computed, inject } from 'vue'
-import { TABLE_INJECTION_KEY } from '../../constants/token.constant'
+import { TABLE_INJECTION_KEY } from '../../shared/constants/token.constant'
 import type { TableColumnCtx } from '../TableColumn/defaults'
 import type { TableHeaderProps } from '.'
