+++ components/table/src/shared/composables/useLayoutObserver.ts
@@ -8,9 +8,9 @@ import {
   onUpdated,
 } from 'vue'
 
-import type { TableHeader } from '../components/TableHeader'
+import type { TableHeader } from '../../components/TableHeader'
 import type TableLayout from './table-layout'
-import type { Table } from '../Table/defaults'
+import type { Table } from '../../Table/defaults'
 
 function useLayoutObserver<T>(root: Table<T>) {
   const instance = getCurrentInstance() as TableHeader
