+++ components/table/src/shared/store/watcher.ts
@@ -15,8 +15,8 @@ import useCurrent from './current'
 import useTree from './tree'
 
 import type { Ref } from 'vue'
-import type { TableColumnCtx } from '../components/TableColumn/defaults'
-import type { DefaultRow, Table, TableRefs } from '../Table/defaults'
+import type { TableColumnCtx } from '../../components/TableColumn/defaults'
+import type { DefaultRow, Table, TableRefs } from '../../Table/defaults'
 import type { StoreFilter } from '.'
 
 const sortData = (data, states) => {
