+++ components/table/src/table/index.vue
@@ -161,8 +161,8 @@ import { Mousewheel } from '@flash-global66/g-utils'
 import { useLocale } from '@flash-global66/g-hooks'
 import { useNamespace } from '@flash-global66/g-utils'
 import GScrollbar from '@flash-global66/g-scrollbar'
-import { createStore } from '../store/helper'
-import TableLayout from '../composables/table-layout'
+import { createStore } from '../shared/store/helper'
+import TableLayout from '../shared/composables/table-layout'
 import TableHeader from '../components/TableHeader'
 import TableBody from '../components/TableBody'
 import TableFooter from '../components/TableFooter'
@@ -171,9 +171,9 @@ import { convertToRows } from '../components/TableHeader/utils-helper'
 import useStyle from './style-helper'
 import useKeyRender from './key-render-helper'
 import defaultProps from './defaults'
-import { TABLE_INJECTION_KEY } from '../constants/token.constant'
-import { hColgroup } from '../utils/render.util'
-import { useScrollbar } from '../composables/useScrollbar'
+import { TABLE_INJECTION_KEY } from '../shared/constants/token.constant'
+import { hColgroup } from '../shared/utils/render.util'
+import { useScrollbar } from '../shared/composables/useScrollbar'
 import { clearExpansionCacheForTable } from '../components/TableColumn/cell-renderers/cell-expansion-utils'
 
 import type { Table } from './defaults'
